Fixed a linux signed/unsigned comparison compile error on line 253.

+++ SimulationCore/branches/DVTE07_Final_Release_Candidate_1_2007_09_10/source/SimCore/Components/ControlStateComponent.cpp
@@ -250,7 +250,7 @@ namespace SimCore
                {
                   // Change the weapon on the associated vehicle if the weapon
                   // selection has changed.
-                  if( info->mWeaponSelected != control->GetCurrentState()
+                  if( info->mWeaponSelected != unsigned(control->GetCurrentState())
                      || ! info->mControlState.valid() )
                   {
                      // Determine if the current control state is the vehicle control
